Express Aula 06.02
Trabalhando com o Controller.
