+++ Express_Desenvolvendo_aplicacoes_web/APP/routes/clients.ts
@@ -4,31 +4,8 @@ import clientsController from '../controllers/clientController';
 const router = Router();
 
 router.get('/', clientsController.index);
-router.get('/saudacao/:nome?', (req, res) =>{
-    if (req.params.nome === undefined){
-        res.send('Desconhecido')
-    } else{
-        res.send(`<h1>Olá ${req.params.nome}</h1>`);
-    }
-})
-router.get('/produto/:id', (req, res) =>{
-    var produtos = [
-        {
-            produto: 'Memória RAM',
-            descricao: '16GB'
-        },
-        {
-            produto: 'Placa de Vídeo',
-            descricao: '8GB'
-        },
-        {
-            produto: 'Monitor',
-            descricao: 'Full HD'
-        }
-    ];
-    res.send(`<h1>${produtos[parseInt(req.params.id)].produto}</h1>
-             <p>${produtos[parseInt(req.params.id)].descricao} </p>`)
-});
+
+router.get('/produto/:id', clientsController.produto);
